Extract out new enforcement API into its own file
