+++ backends/flask-sqlalchemy/app/routes/__init__.py
@@ -1,3 +1,4 @@
+from . import helpers
 from . import issues
 from . import orgs
 from . import repos
